move models_library models where they are actually used
